+++ app/src/components/SpecimenStrip/index.ts
@@ -1,0 +1,2 @@
+export * from './SpecimenStrip';
+export * from './payloads';